Stop using callee-saved %rbx for evaluation
This was the ever elusive bug that has been stopping any real progress
from happening for a long time. The %rbx register is callee-saved, but
that is not respected in code we generate, which crashes spectacularly
when linked to conforming code.

With this simple fix, several more files can be compiled in the
bootstrap assembly without errors.