Rewrite initial line preprocessing
Use fread instead of fgetc, keeping a buffer for preprocessing each file. This
results in a significant performance improvement, most notably in number of
branch instructions executed.

Performance is measured by running bin/lacc src/backend/compile.c, with lacc
compiled with -O3 -DNDEBUG. These numbers are for a previous commit:

    54 910 557      instructions              #    1,55  insns per cycle
    12 283 925      branches                  #  930,206 M/sec
       275 157      branch-misses             #    2,24% of all branches

And with this rewrite:

    48 334 236      instructions              #    1,58  insns per cycle
     9 883 932      branches                  # 1176,186 M/sec
       276 686      branch-misses             #    2,80% of all branches