Introduce ir variable kind for address
In addition to direct, deref and immediate, have an explicit model
of symbol address with offset as variable kind. This reduces the
number of temporary variables generated during evaluation, because
address expressions can be represented directly.

Formally, an ADDRESS var refers to (&sym + offset); closely related
to a DIRECT var, which is *(&sym + offset). We still need DEREF, in
order to model l-values in statements like *a = b.