Replace instruction encoding with table based implementation
With need to support inline assembly, it becomes impractical to
manually encode every instruction separately. Instead store all
relevant information for each instruction in a table, and implement
generic encoding of entries.

This way it will be much easier to add more instructions, and
automatically have them available for the assembler.