The -W flag means -Wextra
Hi --

The -W flag is a (allegedly deprecated) synonym for -Wextra. Both gcc and clang recognize -W.
lacc however will complain that -W is missing an argument. This PR teaches lacc about -W. It still may be the case that -Wextra gets ignored for now, but that's preferable to a compiler error.