Optimize hash table and string implementation
Use open addressing with linear probing instead of a fixed bucket
count with chained lists, and make all string instances globally
unique to enable fast comparison and hashing.

With this is is time to finally update the performance numbers in
the readme, which as of the previous commit were quite misleading.
The actual effect of these optimizations are much more significant
than the difference from earlier measurements, with around 150 M
fewer dynamic instructions executed in the test case.