Introduce configure script
Having just a simple Makefile seemed desirable, but adding this
explicit step makes things more clear. We try to handle machine
types for build, host, and target, and enable cross compilation.
Note that lacc itself is not a cross compiler, so there is no way
to set target different from host.

Test recipes are moved to a separate Makefile, as tests can only be
run when build is native.